@@ -1,16 +1,17 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 28
-    buildToolsVersion '28.0.3'
+    compileSdkVersion 30
+    buildToolsVersion '30.0.3'
 
     defaultConfig {
         applicationId "org.shadowice.flocke.andotp"
-        minSdkVersion 19
-        targetSdkVersion 28
-        versionCode 26
-        versionName "0.6.3.1"
+        minSdkVersion 22
+        targetSdkVersion 30
+        versionCode 35
+        versionName "0.8.0"
         vectorDrawables.useSupportLibrary = true
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
     buildTypes {
         release {
