@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2017-2018 Jakob Nixdorf
+ * Copyright (C) 2017-2020 Jakob Nixdorf
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -27,6 +27,7 @@ import android.app.AlertDialog;
 import android.content.DialogInterface;
 import android.content.Intent;
 import android.content.SharedPreferences;
+import android.net.Uri;
 import android.os.Build;
 import android.os.Bundle;
 import android.preference.CheckBoxPreference;
@@ -36,7 +37,10 @@ import android.preference.Preference;
 import android.preference.PreferenceCategory;
 import android.preference.PreferenceFragment;
 import android.preference.PreferenceManager;
-import android.support.v7.widget.Toolbar;
+import androidx.appcompat.widget.Toolbar;
+
+import android.provider.DocumentsContract;
+import android.util.Log;
 import android.view.ViewStub;
 import android.widget.Toast;
 
@@ -54,6 +58,7 @@ import org.shadowice.flocke.andotp.Utilities.Settings;
 import org.shadowice.flocke.andotp.Utilities.UIHelper;
 
 import java.util.ArrayList;
+import java.util.Locale;
 
 import javax.crypto.SecretKey;
 
@@ -62,10 +67,12 @@ import static org.shadowice.flocke.andotp.Utilities.Constants.EncryptionType;
 
 public class SettingsActivity extends BaseActivity
         implements SharedPreferences.OnSharedPreferenceChangeListener{
-    SettingsFragment fragment;
 
-    SecretKey encryptionKey = null;
-    boolean encryptionChanged = false;
+    private SettingsFragment fragment;
+    private SharedPreferences prefs;
+
+    private SecretKey encryptionKey = null;
+    private boolean encryptionChanged = false;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -85,14 +92,31 @@ public class SettingsActivity extends BaseActivity
         if (keyMaterial != null && keyMaterial.length > 0)
             encryptionKey = EncryptionHelper.generateSymmetricKey(keyMaterial);
 
+        if (savedInstanceState != null) {
+            encryptionChanged = savedInstanceState.getBoolean(Constants.EXTRA_SETTINGS_ENCRYPTION_CHANGED, false);
+
+            byte[] encKey = savedInstanceState.getByteArray(Constants.EXTRA_SETTINGS_ENCRYPTION_KEY);
+            if (encKey != null) {
+                encryptionKey = EncryptionHelper.generateSymmetricKey(encKey);
+            }
+        }
+
         fragment = new SettingsFragment();
 
         getFragmentManager().beginTransaction()
                 .replace(R.id.container_content, fragment)
                 .commit();
 
-        SharedPreferences sharedPref = PreferenceManager.getDefaultSharedPreferences(this);
-        sharedPref.registerOnSharedPreferenceChangeListener(this);
+        prefs = PreferenceManager.getDefaultSharedPreferences(this);
+        prefs.registerOnSharedPreferenceChangeListener(this);
+    }
+
+    @Override
+    protected void onSaveInstanceState(Bundle outState) {
+        super.onSaveInstanceState(outState);
+
+        outState.putBoolean(Constants.EXTRA_SETTINGS_ENCRYPTION_CHANGED, encryptionChanged);
+        outState.putByteArray(Constants.EXTRA_SETTINGS_ENCRYPTION_KEY, encryptionKey.getEncoded());
     }
 
     public void finishWithResult() {
@@ -123,8 +147,9 @@ public class SettingsActivity extends BaseActivity
         backupManager.dataChanged();
 
         if (key.equals(getString(R.string.settings_key_theme)) ||
-                key.equals(getString(R.string.settings_key_locale)) ||
+                key.equals(getString(R.string.settings_key_lang)) ||
                 key.equals(getString(R.string.settings_key_special_features)) ||
+                key.equals(getString(R.string.settings_key_backup_location)) ||
                 key.equals(getString(R.string.settings_key_theme_mode))) {
             recreate();
         } else if(key.equals(getString(R.string.settings_key_encryption))) {
@@ -145,13 +170,16 @@ public class SettingsActivity extends BaseActivity
                 if (fragment.useAndroidSync != null)
                     fragment.useAndroidSync.setEnabled(true);
             }
+        } else if(key.equals(getString(R.string.settings_key_enable_android_backup_service)))
+        {
+            Log.d(SettingsActivity.class.getSimpleName(), "onSharedPreferenceChanged called modifying settings_key_enable_android_backup_service service is now: " +
+                    (settings.getAndroidBackupServiceEnabled() ? "enabled" : "disabled"));
+
+            int message = settings.getAndroidBackupServiceEnabled() ? R.string.settings_toast_android_sync_enabled : R.string.settings_toast_android_sync_disabled;
+            Toast.makeText(this, message, Toast.LENGTH_SHORT).show();
         }
 
-        if (fragment.useAutoBackup != null) {
-            fragment.useAutoBackup.setEnabled(BackupHelper.autoBackupType(this) == Constants.BackupType.ENCRYPTED);
-            if (!fragment.useAutoBackup.isEnabled())
-                fragment.useAutoBackup.setChecked(false);
-        }
+        fragment.updateAutoBackup();
     }
 
     private void generateNewEncryptionKey() {
@@ -216,6 +244,19 @@ public class SettingsActivity extends BaseActivity
         return false;
     }
 
+    private void requestBackupAccess() {
+        Intent intent = new Intent(Intent.ACTION_OPEN_DOCUMENT_TREE);
+        intent.addFlags(Intent.FLAG_GRANT_READ_URI_PERMISSION
+                | Intent.FLAG_GRANT_WRITE_URI_PERMISSION
+                | Intent.FLAG_GRANT_PERSISTABLE_URI_PERMISSION
+                | Intent.FLAG_GRANT_PREFIX_URI_PERMISSION);
+
+        if (android.os.Build.VERSION.SDK_INT >= Build.VERSION_CODES.O && settings.isBackupLocationSet())
+            intent.putExtra(DocumentsContract.EXTRA_INITIAL_URI, settings.getBackupLocation());
+
+        startActivityForResult(intent, Constants.INTENT_SETTINGS_BACKUP_LOCATION);
+    }
+
     @Override
     protected void onActivityResult(int requestCode, int resultCode, Intent data) {
         super.onActivityResult(requestCode, resultCode, data);
@@ -234,24 +275,40 @@ public class SettingsActivity extends BaseActivity
             } else {
                 Toast.makeText(this, R.string.settings_toast_encryption_auth_failed, Toast.LENGTH_LONG).show();
             }
+        } else if (requestCode == Constants.INTENT_SETTINGS_BACKUP_LOCATION && resultCode == RESULT_OK) {
+            Uri treeUri = data.getData();
+            if (treeUri != null) {
+                final int takeFlags = data.getFlags() & (Intent.FLAG_GRANT_READ_URI_PERMISSION | Intent.FLAG_GRANT_WRITE_URI_PERMISSION);
+                getContentResolver().takePersistableUriPermission(treeUri, takeFlags);
+                settings.setBackupLocation(treeUri);
+            }
         } else if (fragment.pgpSigningKey.handleOnActivityResult(requestCode, resultCode, data)) {
             // handled by OpenPgpKeyPreference
             return;
         }
     }
 
+    @Override
+    protected void onDestroy() {
+        prefs.unregisterOnSharedPreferenceChangeListener(this);
+        prefs = null;
+
+        super.onDestroy();
+    }
+
     public static class SettingsFragment extends PreferenceFragment {
-        PreferenceCategory catSecurity;
+        private PreferenceCategory catSecurity;
         PreferenceCategory catUI;
 
-        Settings settings;
-        ListPreference encryption;
-        CheckBoxPreference useAutoBackup;
-        CheckBoxPreference useAndroidSync;
+        private Settings settings;
+        private ListPreference encryption;
+        private Preference backupLocation;
+        private ListPreference useAutoBackup;
+        private CheckBoxPreference useAndroidSync;
 
-        OpenPgpAppPreference pgpProvider;
-        EditTextPreference pgpEncryptionKey;
-        OpenPgpKeyPreference pgpSigningKey;
+        private OpenPgpAppPreference pgpProvider;
+        private EditTextPreference pgpEncryptionKey;
+        private OpenPgpKeyPreference pgpSigningKey;
         ListPreference themeMode;
         ListPreference theme;
 
@@ -277,6 +334,20 @@ public class SettingsActivity extends BaseActivity
                     .show();
         }
 
+        public void updateAutoBackup() {
+            if (useAutoBackup != null) {
+                useAutoBackup.setEnabled(BackupHelper.autoBackupType(getActivity()) == Constants.BackupType.ENCRYPTED);
+                if (!useAutoBackup.isEnabled())
+                    useAutoBackup.setValue(Constants.AutoBackup.OFF.toString().toLowerCase(Locale.ENGLISH));
+
+                if (useAutoBackup.isEnabled()) {
+                    useAutoBackup.setSummary(R.string.settings_desc_auto_backup_password_enc);
+                } else {
+                    useAutoBackup.setSummary(R.string.settings_desc_auto_backup_requirements);
+                }
+            }
+        }
+
         @Override
         public void onCreate(Bundle savedInstanceState) {
             super.onCreate(savedInstanceState);
@@ -294,6 +365,20 @@ public class SettingsActivity extends BaseActivity
                     return ((SettingsActivity) getActivity()).tryEncryptionChange(settings.getEncryption(), newKey);
                 }
             });
+
+            CheckBoxPreference blockAutofill = (CheckBoxPreference) findPreference(getString(R.string.settings_key_block_autofill));
+            CheckBoxPreference autoUnlockAfterAutofill = (CheckBoxPreference) findPreference(getString(R.string.settings_key_auto_unlock_after_autofill));
+            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
+                blockAutofill.setEnabled(true);
+                blockAutofill.setSummary(R.string.settings_desc_block_autofill);
+                autoUnlockAfterAutofill.setEnabled(true);
+                autoUnlockAfterAutofill.setSummary(R.string.settings_desc_auto_unlock_after_autofill);
+            } else {
+                blockAutofill.setEnabled(false);
+                blockAutofill.setSummary(R.string.settings_desc_autofill_requires_android_o);
+                autoUnlockAfterAutofill.setEnabled(false);
+                autoUnlockAfterAutofill.setSummary(R.string.settings_desc_autofill_requires_android_o);
+            }
 
             // Authentication
             catSecurity = (PreferenceCategory) findPreference(getString(R.string.settings_key_cat_security));
@@ -332,6 +417,23 @@ public class SettingsActivity extends BaseActivity
                 }
             });
 
+            // Backup location
+            backupLocation = findPreference(getString(R.string.settings_key_backup_location));
+
+            if (settings.isBackupLocationSet()) {
+                backupLocation.setSummary(R.string.settings_desc_backup_location_set);
+            } else {
+                backupLocation.setSummary(R.string.settings_desc_backup_location);
+            }
+
+            backupLocation.setOnPreferenceClickListener(new Preference.OnPreferenceClickListener() {
+                @Override
+                public boolean onPreferenceClick(Preference preference) {
+                    ((SettingsActivity) getActivity()).requestBackupAccess();
+                    return true;
+                }
+            });
+
             // OpenPGP
             pgpProvider = (OpenPgpAppPreference) findPreference(getString(R.string.settings_key_openpgp_provider));
             pgpEncryptionKey = (EditTextPreference) findPreference(getString(R.string.settings_key_openpgp_key_encrypt));
@@ -360,10 +462,8 @@ public class SettingsActivity extends BaseActivity
                 }
             });
 
-            useAutoBackup = (CheckBoxPreference) findPreference(getString(R.string.settings_key_auto_backup_password_enc));
-            useAutoBackup.setEnabled(BackupHelper.autoBackupType(getActivity()) == Constants.BackupType.ENCRYPTED);
-            if(!useAutoBackup.isEnabled())
-                useAutoBackup.setChecked(false);
+            useAutoBackup = (ListPreference)findPreference(getString(R.string.settings_key_auto_backup_password_enc));
+            updateAutoBackup();
 
             useAndroidSync = (CheckBoxPreference) findPreference(getString(R.string.settings_key_enable_android_backup_service));
             useAndroidSync.setEnabled(settings.getEncryption() == EncryptionType.PASSWORD);
