@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2017-2018 Jakob Nixdorf
+ * Copyright (C) 2017-2020 Jakob Nixdorf
  * Copyright (C) 2017 Carlos Melero
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
