@@ -1,6 +1,6 @@
 /*
- * Copyright (C) 2018 Jakob Nixdorf
- * Copyright (C) 2018 Richy HBM
+ * Copyright (C) 2018-2020 Jakob Nixdorf
+ * Copyright (C) 2018-2020 Richy HBM
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -25,7 +25,6 @@ package org.shadowice.flocke.andotp.Receivers;
 
 import android.content.Context;
 import android.content.Intent;
-import android.net.Uri;
 
 import org.shadowice.flocke.andotp.Database.Entry;
 import org.shadowice.flocke.andotp.R;
@@ -56,8 +55,6 @@ public class EncryptedBackupBroadcastReceiver extends BackupBroadcastReceiver {
             if (!canSaveBackup(context))
                 return;
 
-            Uri savePath = Tools.buildUri(settings.getBackupDir(), BackupHelper.backupFilename(context, Constants.BackupType.ENCRYPTED));
-
             String password = settings.getBackupPasswordEnc();
 
             if (password.isEmpty()) {
@@ -65,7 +62,7 @@ public class EncryptedBackupBroadcastReceiver extends BackupBroadcastReceiver {
                 return;
             }
 
-            SecretKey encryptionKey = null;
+            SecretKey encryptionKey;
 
             if (settings.getEncryption() == Constants.EncryptionType.KEYSTORE) {
                 encryptionKey = KeyStoreHelper.loadEncryptionKeyFromKeyStore(context, false);
@@ -75,6 +72,13 @@ public class EncryptedBackupBroadcastReceiver extends BackupBroadcastReceiver {
             }
 
             if (Tools.isExternalStorageWritable()) {
+                BackupHelper.BackupFile cryptBackupFile = BackupHelper.backupFile(context, settings.getBackupLocation(), Constants.BackupType.ENCRYPTED);
+
+                if (cryptBackupFile.file == null) {
+                    NotificationHelper.notify(context, Constants.NotificationChannel.BACKUP_FAILED, R.string.backup_receiver_title_backup_failed, cryptBackupFile.errorMessage);
+                    return;
+                }
+
                 ArrayList<Entry> entries = DatabaseHelper.loadDatabase(context, encryptionKey);
                 String plain = DatabaseHelper.entriesToString(entries);
 
@@ -92,9 +96,9 @@ public class EncryptedBackupBroadcastReceiver extends BackupBroadcastReceiver {
                     System.arraycopy(salt, 0, data, Constants.INT_LENGTH, Constants.ENCRYPTION_IV_LENGTH);
                     System.arraycopy(encrypted, 0, data, Constants.INT_LENGTH + Constants.ENCRYPTION_IV_LENGTH, encrypted.length);
 
-                    StorageAccessHelper.saveFile(context, savePath, data);
+                    StorageAccessHelper.saveFile(context, cryptBackupFile.file.getUri(), data);
 
-                    NotificationHelper.notify(context, Constants.NotificationChannel.BACKUP_SUCCESS, R.string.backup_receiver_title_backup_success, savePath.getPath());
+                    NotificationHelper.notify(context, Constants.NotificationChannel.BACKUP_SUCCESS, R.string.backup_receiver_title_backup_success, cryptBackupFile.file.getName());
                 } catch (Exception e) {
                     e.printStackTrace();
                     NotificationHelper.notify(context, Constants.NotificationChannel.BACKUP_FAILED, R.string.backup_receiver_title_backup_failed, R.string.backup_toast_export_failed);
